+++ src/TelegramLogChannelServiceProvider.php
@@ -43,6 +43,7 @@ class TelegramLogChannelServiceProvider extends ServiceProvider
                 'token' => env('TELEGRAM_LOG_BOT_TOKEN'),
                 'chat_id' => env('TELEGRAM_LOG_CHAT_ID'),
                 'level' => env('TELEGRAM_LOG_LEVEL', 'error'),
+                'throttle' => env('TELEGRAM_LOG_THROTTLE', 600),
             ]);
         }